로그인시 학습기록 작성하도록 구현

+++ hantalk/src/main/java/com/example/hantalk/dto/ResourceDTO.java
@@ -8,7 +8,7 @@ import java.time.LocalDateTime;
 @Getter
 @Setter
 public class ResourceDTO {
-    private Long resourceId;
+    private int resourceId;
     private String title;
     private String content;
     private String archive;
